Yet another 64-bit tweak like all the others... I'm too lazy to post it and wait for votes.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@79706 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/http_config.c
@@ -876,7 +876,7 @@ const char *set_file_slot(cmd_parms *cmd, char *struct_ptr, char *arg)
        This allows .htaccess to be independent of server_root,
        so the server can be moved or mirrored with less pain.  */
     char *p;
-    int offset = (int) cmd->info;
+    int offset = (int) (long) cmd->info;
     if (os_is_path_absolute(arg))
 	p = pstrdup(cmd->pool, arg);
     else
